KP-8569 Add mechanism for tweaking CMDI, trim self-link
This allows easily adding new modifications to the CMDI metadata while
hopefully keeping the code readable-ish. The first modification is to
make the self-link not be the full link
(http://urn.fi/urn:nbn:Bfi:lb-123) but instead just the URN
(urn:nbn:fi:lb-123). This works in COMEDI and is neater.